xilinx-bootbin: check if Bitstream file exists while do_deploy
Issue: LINUXEXEC-37078

When generating BOOT.BIN in "Full Design" for zynq/zynqmp, there is no
Bitstream file to be integrated into BOOT.BIN. So we need to add a
condition check for this.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

+++ meta-xilinx-core/recipes-bsp/bootbin/xilinx-bootbin_1.0.bb
@@ -265,11 +265,15 @@ do_deploy() {
 }
 
 do_deploy:append:zynqmp () {
-    install -m 0644 ${B}/*.bit ${DEPLOYDIR}/boot.bin-extracted/.
+    if [ -f ${B}/*.bit ]; then
+		install -m 0644 ${B}/*.bit ${DEPLOYDIR}/boot.bin-extracted/.
+    fi
 }
 
 do_deploy:append:zynq () {
-    install -m 0644 ${B}/*.bit ${DEPLOYDIR}/boot.bin-extracted/.
+    if [ -f ${B}/*.bit ]; then
+		install -m 0644 ${B}/*.bit ${DEPLOYDIR}/boot.bin-extracted/.
+    fi
 }
 
 do_deploy:append:versal () {
